+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
 ## [Unreleased]
-- Animated comparison plot.
 - Bubble velocity tracker.
 
 ## [0.1.0]
 ### Added
 - Static module.
 - Plasma density profile plotter in static module.
+- Load number density from file for animated plots.
+- Animated comparison plot.